+++ vdso/filesystem.c
@@ -168,6 +168,6 @@ EXPORT int hydrogen_fs_pipe(int fds[2], int flags) {
     return 0;
 }
 
-EXPORT hydrogen_ret_t hydrogen_fs_ioctl(int file, int request, void *buffer, size_t size) {
+EXPORT hydrogen_ret_t hydrogen_fs_ioctl(int file, unsigned long request, void *buffer, size_t size) {
     return SYSCALL4(SYSCALL_FS_IOCTL, file, request, buffer, size);
 }
